+++ lib/data/remote/entity/response_entity.dart
@@ -1,0 +1,18 @@
+///
+final class ResponseEntity {
+  ///
+  ResponseEntity({
+    required this.body,
+    required this.request,
+    required this.statusCode,
+  });
+
+  ///
+  final String body;
+
+  ///
+  final String request;
+
+  ///
+  final int statusCode;
+}